requirements.txt
Flask
transformers
torch
together
python-docx
PyMuPDF
python-dotenv
